backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
web3==6.11.0
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.26.3
pandas==2.2.1
scipy==1.12.0
tensorflow==2.15.0
alchemy-sdk==3.0.0
python-dotenv==1.0.0
eth-account==0.10.0
openai==1.7.2
requests==2.31.0
eth-utils==2.3.0
eth-abi==4.2.1
eth-typing==3.5.0
eth-keyfile==0.6.1
eth-rlp==0.3.0
rlp==3.0.0
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.2.0
psycopg2-binary==2.9.9
